Rename `thing_property` to `properties`
The module that includes `property` and `setting` is renamed.
There are two main reasons for this:
1. The "thing" in the name is a bit redundant, and it matches "actions".
2. `lt.thing_property` is a symbol we've removed in this branch. Using it as a module name makes for confusing error messages.
