+++ src/labthings_fastapi/properties.py
